Move unbounded list into its own module
